working on on off map
