+++ api/mod_onoff/views.py
@@ -1,6 +1,8 @@
 import os
 import sys
 import time
+import json
+from decimal import Decimal
 
 from flask import make_response, Blueprint, redirect
 from flask import url_for,render_template, jsonify, request
@@ -11,7 +13,7 @@ from helper import Helper
 from spatial import Spatial
 from api import app, db
 from api import debug, error, Session
-
+from ..shared.helper import Helper as h
 
 STATIC_DIR = '/onoff'
 mod_onoff = Blueprint('onoff', __name__, url_prefix='/onoff')
@@ -158,11 +160,77 @@ def surveyor_summary_query():
     debug(response)
     return jsonify(users=response)
 
-#@mod_onoff.route('/map')
-#def map():
-#    return render_template(static('map.html'))
+@mod_onoff.route('/map')
+def map():
+    routes = [ route['rte_desc'] for route in h.get_routes() ]
+    directions = h.get_directions()
+    return render_template(static('map.html'),
+        routes=routes, directions=directions
+    )
 
+@mod_onoff.route('/map/_details', methods=['GET'])
+def map_details():
+    response = {'success':False}
+    if 'rte_desc' in request.args.keys():
+        rte_desc = request.args['rte_desc'].strip()
+        rte = h.rte_lookup(rte_desc)
+        session = Session()
+        fields = ['dir', 'tad', 'centroid', 'stops', 'ons', 'count']
+        query = session.execute("""
+            SELECT """ + ','.join(fields) + """
+            FROM long.tad_stats
+            WHERE rte = :rte;""", {'rte':rte})
+        query_time = session.execute("""
+            SELECT """ + ','.join(fields) + """, bucket
+            FROM long.tad_time_stats
+            WHERE rte = :rte;""", {'rte':rte})
+        query_routes = session.execute("""
+            SELECT dir, ST_AsGeoJson(ST_Transform(ST_Union(geom), 4326))
+            FROM tm_routes
+            WHERE rte = :rte
+            GROUP BY dir;""", {'rte':rte})
+        
+        def build_data(record):
+            data = {}
+            for index in range(1, len(fields)):
+                field = record[index]
+                if isinstance(field, Decimal): field = int(field)
+                data[fields[index]] = field
+            return data
 
+        def int_zero(value):
+            try:
+                return int(value)
+            except:
+                return False
 
+        data = {}
+        data[0] = []
+        data[1] = []
+        for record in query:
+            data[record[0]].append(build_data(record))
+        time_data = {}
+        time_data[0] = {}
+        time_data[1] = {}
+        for record in query_time:
+            tad = record[1]
+            if tad not in time_data[record[0]]:
+                time_data[record[0]][tad] = []
+            ons = int_zero(record[4])
+            count = int_zero(record[5])
+            bucket = int_zero(record[6])
+            time_data[record[0]][tad].insert(bucket, {"count":count, "ons":ons})
+        routes_geom = {}
+        for record in query_routes:
+            routes_geom[record[0]] = {
+                'dir':record[0],
+                'geom':json.loads(record[1])
+            }
+        response['success'] = True
+        response['data'] = data
+        response['time_data'] = time_data
+        response['routes'] = routes_geom
+        session.close()
+    return jsonify(response)
