+++ api/templates/onoff/base.html
@@ -9,6 +9,7 @@
 <li><a href="{{ url_for('onoff.surveyor_status') }}">Surveyors</a></li>
 <li><a href="{{ url_for('onoff.status') }}">Status</a></li>
 <li><a href="{{ url_for('onoff.data') }}">Data</a></li>
+<li><a href="{{ url_for('onoff.map') }}">Map</a></li>
 <!-- need to add brackets another bracket around url_for() if enabling long -->
 <!--li><a href="{ url_for('long.index') }">Intercept</a></li-->
 {% endblock %}
